Move action db models to individual files
