+++ st2actions/st2actions/resultstracker.py
@@ -23,7 +23,7 @@ from oslo.config import cfg
 
 from st2actions.query.base import QueryContext
 from st2common import log as logging
-from st2common.models.db import action as action_models
+from st2common.models.db.executionstate import ActionExecutionStateDB
 from st2common.persistence.executionstate import ActionExecutionState
 from st2common.transport import actionexecutionstate, consumers, publishers
 
@@ -35,7 +35,7 @@ ACTIONSTATE_WORK_Q = actionexecutionstate.get_queue('st2.resultstracker.work',
 
 
 class ResultsTracker(consumers.MessageHandler):
-    message_type = action_models.ActionExecutionStateDB
+    message_type = ActionExecutionStateDB
 
     def __init__(self, connection, queues):
         super(ResultsTracker, self).__init__(connection, queues)
