+++ st2actions/st2actions/scheduler.py
@@ -19,7 +19,7 @@ from oslo.config import cfg
 from st2common import log as logging
 from st2common.constants import action as action_constants
 from st2common.exceptions.db import StackStormDBObjectNotFoundError
-from st2common.models.db import action as action_models
+from st2common.models.db.liveaction import LiveActionDB
 from st2common.services import executions
 from st2common.persistence.liveaction import LiveAction
 from st2common.transport import consumers, liveaction
@@ -33,7 +33,7 @@ ACTIONRUNNER_REQUEST_Q = liveaction.get_status_management_queue(
 
 
 class ActionExecutionScheduler(consumers.MessageHandler):
-    message_type = action_models.LiveActionDB
+    message_type = LiveActionDB
 
     def process(self, request):
         """Schedules the LiveAction and publishes the request
@@ -42,7 +42,7 @@ class ActionExecutionScheduler(consumers.MessageHandler):
         LiveAction in statuses other than "requested" are ignored.
 
         :param request: Action execution request.
-        :type request: ``st2common.models.db.action.LiveActionDB``
+        :type request: ``st2common.models.db.liveaction.LiveActionDB``
         """
 
         if request.status != action_constants.LIVEACTION_STATUS_REQUESTED:
