+++ st2actions/tests/unit/test_actionchain.py
@@ -22,7 +22,7 @@ from st2common.constants.action import LIVEACTION_STATUS_RUNNING
 from st2common.constants.action import LIVEACTION_STATUS_SUCCEEDED
 from st2common.constants.action import LIVEACTION_STATUS_FAILED
 from st2common.models.api.notification import NotificationsHelper
-from st2common.models.db.action import LiveActionDB
+from st2common.models.db.liveaction import LiveActionDB
 from st2common.models.db.datastore import KeyValuePairDB
 from st2common.models.system.common import ResourceReference
 from st2common.persistence.datastore import KeyValuePair
