+++ st2actions/tests/unit/test_mistral_v2.py
@@ -39,7 +39,7 @@ from st2actions.runners.mistral.v2 import MistralRunner
 from st2common.constants import action as action_constants
 from st2common.models.api.access import TokenAPI
 from st2common.models.api.action import ActionAPI
-from st2common.models.db.action import LiveActionDB
+from st2common.models.db.liveaction import LiveActionDB
 from st2common.persistence.action import Action
 from st2common.persistence.liveaction import LiveAction
 from st2common.services import access as access_service
