+++ st2actions/tests/unit/test_queue_consumers.py
@@ -23,7 +23,7 @@ from kombu.message import Message
 from st2actions import scheduler, worker
 from st2actions.container.base import RunnerContainer
 from st2common.constants import action as action_constants
-from st2common.models.db.action import LiveActionDB
+from st2common.models.db.liveaction import LiveActionDB
 from st2common.models.system.common import ResourceReference
 from st2common.persistence import action
 from st2common.services import executions
