+++ st2api/st2api/controllers/exp/aliasexecution.py
@@ -22,7 +22,8 @@ from pecan import rest
 from st2common import log as logging
 from st2common.models.api.base import jsexpose
 from st2common.models.api.action import AliasExecutionAPI
-from st2common.models.db.action import LiveActionDB, NotificationSchema, NotificationSubSchema
+from st2common.models.db.liveaction import LiveActionDB
+from st2common.models.db.notification import NotificationSchema, NotificationSubSchema
 from st2common.models.utils import action_alias_utils, action_param_utils
 from st2common.persistence.actionalias import ActionAlias
 from st2common.services import action as action_service
