+++ st2common/st2common/models/api/action.py
@@ -19,9 +19,11 @@ from st2common import log as logging
 from st2common.models.api.base import BaseAPI
 from st2common.models.api.tag import TagsHelper
 from st2common.models.api.notification import (NotificationSubSchemaAPI, NotificationsHelper)
-from st2common.models.db.action import (RunnerTypeDB, ActionDB, LiveActionDB)
-from st2common.models.db.action import ActionExecutionStateDB
-from st2common.models.db.action import ActionAliasDB
+from st2common.models.db.action import ActionDB
+from st2common.models.db.actionalias import ActionAliasDB
+from st2common.models.db.executionstate import ActionExecutionStateDB
+from st2common.models.db.liveaction import LiveActionDB
+from st2common.models.db.runner import RunnerTypeDB
 from st2common.constants.action import LIVEACTION_STATUSES
 from st2common.models.system.common import ResourceReference
