+++ st2common/st2common/models/db/__init__.py
@@ -14,10 +14,14 @@ LOG = logging.getLogger(__name__)
 MODEL_MODULE_NAMES = [
     'st2common.models.db.access',
     'st2common.models.db.action',
-    'st2common.models.db.actionrunner',
+    'st2common.models.db.actionalias',
     'st2common.models.db.datastore',
     'st2common.models.db.execution',
+    'st2common.models.db.executionstate',
+    'st2common.models.db.liveaction',
+    'st2common.models.db.policy',
     'st2common.models.db.rule',
+    'st2common.models.db.runner',
     'st2common.models.db.sensor',
     'st2common.models.db.trigger',
 ]
