+++ st2common/tests/unit/services/test_action.py
@@ -18,7 +18,7 @@ import jsonschema
 
 from st2actions.container.base import RunnerContainer
 from st2common.constants import action as action_constants
-from st2common.models.db.action import LiveActionDB
+from st2common.models.db.liveaction import LiveActionDB
 from st2common.models.api.action import RunnerTypeAPI, ActionAPI
 from st2common.models.system.common import ResourceReference
 from st2common.persistence.action import Action
