+++ st2common/tests/unit/test_action_db_utils.py
@@ -22,7 +22,8 @@ import mock
 from st2common.exceptions.db import StackStormDBObjectNotFoundError
 from st2common.transport.publishers import PoolPublisher
 from st2common.models.api.action import RunnerTypeAPI
-from st2common.models.db.action import (ActionDB, LiveActionDB)
+from st2common.models.db.action import ActionDB
+from st2common.models.db.liveaction import LiveActionDB
 from st2common.models.system.common import ResourceReference
 from st2common.persistence.action import Action
 from st2common.persistence.liveaction import LiveAction
