+++ st2common/tests/unit/test_db_action_state.py
@@ -1,6 +1,6 @@
 import bson
 
-from st2common.models.db.action import ActionExecutionStateDB
+from st2common.models.db.executionstate import ActionExecutionStateDB
 from st2common.persistence.executionstate import ActionExecutionState
 
 from st2tests import DbTestCase
