+++ st2common/tests/unit/test_db_liveaction.py
@@ -17,7 +17,8 @@ import datetime
 
 import mock
 
-from st2common.models.db.action import LiveActionDB, NotificationSchema, NotificationSubSchema
+from st2common.models.db.liveaction import LiveActionDB
+from st2common.models.db.notification import NotificationSchema, NotificationSubSchema
 from st2common.persistence.liveaction import LiveAction
 from st2common.transport.publishers import PoolPublisher
 from st2common.util import isotime
